% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demo_historique.R
\name{champ_insee}
\alias{champ_insee}
\title{Champ géographique des séries de l'Insee}
\usage{
champ_insee(annee)
}
\arguments{
\item{annee}{Années.}
}
\value{
Un vecteur de libellés (\code{NA} pour les années 1915 à 1919, non
couvertes par les séries).
}
\description{
Renvoie, pour chaque année, le champ géographique des séries de
population de l'Insee :
}
\details{
\itemize{
\item \code{"metropole_hors_alsace_moselle"} : 1901-1914 (frontières de 1871) et
1939-1945, sauf 1944 ;
\item \code{"metropole_hors_alsace_moselle_corse"} : 1944 ;
\item \code{"metropole"} : 1920-1938 et 1946-1994 ;
\item \code{"france_hors_mayotte"} : 1995-2013 ;
\item \code{"france"} : à partir de 2014.
}
}
\examples{
champ_insee(c(1910, 1944, 1980, 2000, 2030))
}
