% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sources.R
\name{url_source}
\alias{url_source}
\title{Adresse d'un fichier source}
\usage{
url_source(objet, scenario = "central", millesime = NULL)
}
\arguments{
\item{objet}{Type de données (voir \code{\link[=sources_donnees]{sources_donnees()}}).}

\item{scenario}{Scénario (par défaut \code{"central"}).}

\item{millesime}{Millésime ; \code{NULL} pour le plus récent.}
}
\value{
Une chaîne de caractères.
}
\description{
Renvoie l'adresse du fichier correspondant à un type de données, un
scénario et un millésime. Par défaut, le millésime le plus récent du
registre est retenu.
}
\examples{
url_source("projpop")
}
