% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/telechargement.R
\name{vider_cache}
\alias{vider_cache}
\title{Vider le cache des fichiers téléchargés}
\usage{
vider_cache()
}
\value{
\code{TRUE} de manière invisible.
}
\description{
Vider le cache des fichiers téléchargés
}
