% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pilotretrtools-package.R
\docType{package}
\name{pilotretrtools-package}
\alias{pilotretrtools}
\alias{pilotretrtools-package}
\title{pilotretrtools : outils pour illustrer le pilotage d'un régime par annuités}
\description{
Le package est organisé en familles de fonctions, repérables par leur
préfixe :
}
\details{
\itemize{
\item \verb{lire_*()} : lecture des fichiers diffusés (Insee, COR, barèmes IPP) ;
\item \verb{prolonger_*()} : prolongation des projections au-delà de leur horizon ;
\item \verb{construire_*()} : assemblage des tables d'hypothèses ;
\item \verb{calculer_*()} et \verb{decomposer_*()} : calculs du modèle ;
\item \verb{indic_*()} : indicateurs par génération ou par année ;
\item \verb{graph_*()} et \verb{donnees_graph_*()} : graphiques et données sous-jacentes ;
\item \verb{app_*()} : applications Shiny.
}

Les conventions communes (grille génération x âge x année, définition des
âges) sont décrites dans \code{vignette("conventions", package = "pilotretrtools")}
et vérifiées par \code{\link[=valider_grille]{valider_grille()}}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/patrickaubert/pilotretrtools}
  \item Report bugs at \url{https://github.com/patrickaubert/pilotretrtools/issues}
}

}
\author{
\strong{Maintainer}: Patrick Aubert \email{patrick.aubert@ipp.eu}

}
\keyword{internal}
